Refactor data transformation logic into a separate `transformData` function
