+++ src/functions/index.ts
@@ -3,6 +3,7 @@ export { cloneDeep } from './clone-deep'
 export { getCollatorComparator } from './get-collator-comparator'
 export { load } from './load'
 export { save } from './save'
+export { transformData } from './transform-data'
 export { sortCompare } from './sort-compare'
 export { warnDeprecated } from './warn-deprecated'
 export { warnInfo } from './warn-info'
